frontend move main.dart back to default location for convenience
