+++ gemini_proact_flutter/test/widget_test.dart
@@ -8,7 +8,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:gemini_proact_flutter/view/main.dart';
+import 'package:gemini_proact_flutter/view/root.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
